Refactor: rename translate service functions
