+++ api/lib/domain/usecases/get-learning-content-for-replication.js
@@ -7,7 +7,7 @@ import {
 } from '../../infrastructure/datasources/airtable/index.js';
 import { challengeRepository, localizedChallengeRepository, competenceRepository, skillRepository } from '../../infrastructure/repositories/index.js';
 import { knex } from '../../../db/knex-database-connection.js';
-import { translateChallenges } from '../services/translate-challenges.js';
+import { createChallengeTranslator } from '../services/translate-challenge.js';
 
 export async function getLearningContentForReplication() {
   const [
@@ -33,9 +33,9 @@ export async function getLearningContentForReplication() {
     thematicDatasource.list(),
     _getCoursesFromPGForReplication(),
   ]);
-  const translate = translateChallenges({ localizedChallenges });
+  const translateChallenge = createChallengeTranslator({ localizedChallenges });
 
-  const translatedChallenges = challenges.flatMap((challenge) => translate(challenge)).map(normalizeChallenge);
+  const translatedChallenges = challenges.flatMap(translateChallenge).map(normalizeChallenge);
 
   return {
     areas,
